// client/src/setup/store/actionsTypes/survey.actionTypes.ts
export enum SurveyActionTypes {
  REQUEST_CREATE_SURVEY = "REQUEST_CREATE_SURVEY",
  REQUEST_LOAD_SURVEYS = "REQUEST_LOAD_SURVEYS",
  REQUEST_UPDATE_SURVEY = "REQUEST_UPDATE_SURVEY",
  REQUEST_DELETE_SURVEY = "REQUEST_DELETE_SURVEY",
  REQUEST_GET_RECENT_SURVEYS = "REQUEST_GET_RECENT_SURVEYS",
  REQUEST_GET_AVAILABLE_SURVEYS = "REQUEST_GET_AVAILABLE_SURVEYS",
  SUCCESS_LOAD_SURVEYS = "SUCCESS_LOAD_SURVEYS",
  SUCCESS_GET_RECENT_SURVEYS = "SUCCESS_GET_RECENT_SURVEYS",
  SUCCESS_GET_AVAILABLE_SURVEYS = "SUCCESS_GET_AVAILABLE_SURVEYS",
  SUCCESS_UPDATE_SURVEY = "SUCCESS_UPDATE_SURVEY",
  SUCCESS_DELETE_SURVEY = "SUCCESS_DELETE_SURVEY",
  SUCCESS_CREATE_SURVEY = "SUCCESS_CREATE_SURVEY",
  FAILURE_CREATE_SURVEY = "CREATE_SURVEY_FAILURE",
  FAILURE_LOAD_SURVEYS = "FAILURE_LOAD_SURVEYS",
  FAILURE_GET_RECENT_SURVEYS = "FAILURE_GET_RECENT_SURVEYS",
  FAILURE_GET_AVAILABLE_SURVEYS = "FAILURE_GET_AVAILABLE_SURVEYS",
  FAILURE_UPDATE_SURVEY = "FAILURE_UPDATE_SURVEY",
  FAILURE_DELETE_SURVEY = "FAILURE_DELETE_SURVEY",
  REQUEST_GET_GLOBAL_SURVEY_INFO = "REQUEST_GET_GLOBAL_SURVEY_INFO",
  SUCCESS_GET_GLOBAL_SURVEY_INFO = "SUCCESS_GET_GLOBAL_SURVEY_INFO",
  FAILURE_GET_GLOBAL_SURVEY_INFO = "FAILURE_GET_GLOBAL_SURVEY_INFO",
}